Conflicts:
	krita/image/kis_strokes_queue.cpp
	krita/image/tests/kis_strokes_queue_test.cpp
	krita/image/tests/kis_strokes_queue_test.h
	krita/ui/kis_view2.cpp

+++ krita/plugins/filters/blur/kis_wdg_gaussian_blur.h
@@ -39,8 +39,8 @@ public:
     virtual KisPropertiesConfiguration* configuration() const;
 
 private slots:
-    void horizontalRadiusChanged(int);
-    void verticalRadiusChanged(int);
+    void horizontalRadiusChanged(qreal);
+    void verticalRadiusChanged(qreal);
     void aspectLockChanged(bool);
 
 private:
